add retrieval result visualization (#17)
colab:
https://colab.research.google.com/drive/107JN9AqBcd-7m9TJ2OBlMX7H0wTA-MNu?usp=sharing
Certain "incorrect" pages can be considered valid results as they
precisely match the information described in the query. This occurs with
queries that describe commonly depicted scenes, see example query "A
girl hugs a boy."
In some cases, incorrect retrieval results do include the objects
mentioned in the query, but they may be missing other crucial elements.
For instance, an incorrect retrieval of the query "A man points a gun at
his own head" includes image with a gun present, but the gun is not
directly aimed at the character's head.